Fix that non-existing route like '/api/bookss/' redirect to login in API REST instead of returning a 404 NOT_FOUND. Also, fix DELETE /api/books/:id , which always return 200 OK, also when book doesn't exist

+++ bloque_4/spring_security/security_ejem10/src/main/java/es/codeurjc/daw/library/controller/BookRestController.java
@@ -70,11 +70,10 @@ public class BookRestController {
 	@DeleteMapping("/{id}")
 	public ResponseEntity<Book> deleteBook(@PathVariable long id) {
 
-		try {
+		if(service.exist(id)) {
 			service.delete(id);
 			return new ResponseEntity<>(null, HttpStatus.OK);
-
-		} catch (EmptyResultDataAccessException e) {
+		} else {
 			return new ResponseEntity<>(null, HttpStatus.NOT_FOUND);
 		}
 	}
